requirements.txt
python-dotenv  
praw
pymongo
dnspython
jsonschema